mmc: core: Resume SD bus in mmc_sd_detect
If mmc driver calls mmc_get_card(), it gets the claim-host lock
and also resumes the bus if it is in the suspended state.
But if the driver uses mmc_try_claim_host() to get the host lock,
it has to explicitly resume the bus if the bus is in the suspended
state.

So make explicit deferred resume call where-ever driver uses
mmc_try_claim_host() to get the lock.

Change-Id: I47ea61af0b06aa43eaf00c01b3467cec106f4957
Signed-off-by: Veerabhadrarao Badiganti <vbadigan@codeaurora.org>